Move plugins to src root directory.
