+++ npm-packages/meteor-vite/src/meteor/IPC/methods/build-production.ts
@@ -1,9 +1,9 @@
 import { RollupOutput, RollupWatcher } from 'rollup';
 import { build, InlineConfig, resolveConfig } from 'vite';
 import { MeteorViteConfig } from '../../../MeteorViteConfig';
-import { MeteorStubs } from '../../../vite';
+import { MeteorStubs } from '../../../plugin';
 import MeteorVitePackage from '../../../../package.json';
-import { PluginSettings, ProjectJson } from '../../../vite/plugin/MeteorStubs';
+import { PluginSettings, ProjectJson } from '../../../plugin/stubs/MeteorStubs';
 import CreateIPCInterface, { IPCReply } from '../interface';
 
 export default CreateIPCInterface({
