+++ npm-packages/meteor-vite/src/plugin/index.ts
@@ -1,0 +1,1 @@
+export { MeteorStubs } from './stubs/MeteorStubs';
