+++ npm-packages/meteor-vite/src/plugin/stubs/MeteorStubs.ts
