--- npm-packages/meteor-vite/src/vite/index.ts
@@ -1,1 +1,0 @@
-export { MeteorStubs } from './plugin/MeteorStubs';
